Cleaned up index.html; moved scripts into associated .ts and .css files respectively
